refactor: move function to rulesUtils
Added a new one which returns a transferable dimension from a specific
axis.
